ok_dernier_test

+++ project/v1/includes/utils_tmp.hpp
@@ -27,7 +27,7 @@ public:
 	static std::string extract_body(std::string &buff);
 	static int hexa_to_dec(const char *hexVal);
 	static void remove_return(std::string &str);
-	static int	getSecondsDiff(std::string complete_time);
+	static size_t getSecondsDiff(std::string complete_time);
 	static void	ft_getline(std::string &b, std::string &line);
 	static std::vector<std::string> split(const std::string& str, char delim);
 	static std::string trim(const std::string& str);
